Fix hard-coded feature extractor unpacking
Feature extractor classes to be used in `audioanalyser.py` are
enumerated as a tuple at the top of the file. They are instantiated
when processing begins by looping through the tuple and populating
self._featureExtractors with new instances of each class.

Previously, in the `_pickThumbnail` method, we used them by
hard-coded tuple unpacking, but this is prone to error if the actual
feature extractors class tuple ever changes.

- Now, self._featureExtractors is a dictionary of form
`fe_class: fe_instance`. In this way, we can call on each feature
extractor explicitly, by looking up the instance related to an actual
class.
- This does mean we can only have one instance of each type of feature
extractor.